refactor: extract handling of unary expr
